It's 3am, gimmie a break... ugh.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C23960

+++ Rawr.Base/Wowhead.cs
@@ -158,11 +158,11 @@ namespace Rawr
 					break;
 
 				case "dmgmin1":
-					item.MinDamage += int.Parse(value);
+					item.MinDamage += (int)float.Parse(value);
 					break;
 
 				case "dmgmax1":
-					item.MaxDamage += int.Parse(value);
+					item.MaxDamage += (int)float.Parse(value);
 					break;
 
 				case "dmgtype1":
